+++ examples/nanoChatGPT/shared.py
@@ -32,21 +32,6 @@ def setup(config):
     return ctx
 
 
-def create_infinite_dataloader(data, config, collate_fn):
-    """
-    Creates a dataloader and yields batches from it indefinitely, so that we can request
-    batches whenever we like with next.
-    """
-    dl = DataLoader(
-        data,
-        batch_size=config["batch_size"],
-        shuffle=True,  # TODO: perhaps validation set shouldn't be shuffled?
-        collate_fn=collate_fn,
-        drop_last=True,
-    )
-    while True:
-        yield from dl
-
 
 def create_lr_scheduler(config):
     # learning rate decay scheduler (cosine with warmup)
